Fleshed out url config
The `'url'` option for li3_docs only affected a small amount
of the routing. It now properly works for the breadcrumb urls
and search. Also updated the image urls in the css to be
relative so this can be placed in a different location.

Updated the README with a couple examples of how the url option
can be used.